MAT-365: Refactor, move CreateDonation function from controller to new Service class
