+++ src/Domain/DomainException/CampaignNotOpen.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MatchBot\Domain\DomainException;
+
+class CampaignNotOpen extends DomainException
+{
+}
