+++ src/Domain/DomainException/DonationCreateModelLoadFailure.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MatchBot\Domain\DomainException;
+
+class DonationCreateModelLoadFailure extends DomainException
+{
+}
