+++ src/Domain/DomainException/StripeAccountIdNotSetForAccount.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MatchBot\Domain\DomainException;
+
+class StripeAccountIdNotSetForAccount extends DomainException
+{
+}
